TEIID-5596 further refinements to the infinispan separation
