+++ olingo/src/test/java/org/teiid/olingo/TestODataIntegration.java
@@ -59,7 +59,6 @@ import org.teiid.adminapi.Admin.SchemaObjectType;
 import org.teiid.adminapi.CacheStatistics;
 import org.teiid.adminapi.Model;
 import org.teiid.adminapi.impl.ModelMetaData;
-import org.teiid.cache.infinispan.InfinispanCacheFactory;
 import org.teiid.core.TeiidRuntimeException;
 import org.teiid.core.util.ObjectConverterUtil;
 import org.teiid.core.util.TimestampWithTimezone;
@@ -215,7 +214,6 @@ public class TestODataIntegration {
         teiid = new EmbeddedServer();
         EmbeddedConfiguration config = new EmbeddedConfiguration();
         config.setTransactionManager(new DummyBaseTransactionManager());
-        config.setCacheFactory(InfinispanCacheFactory.buildCache(null, config.getTransactionManager()));
         SocketConfiguration sc = new SocketConfiguration();
         sc.setBindAddress("localhost");
         sc.setPortNumber(31000);
