+++ runtime/src/main/resources/org/teiid/runtime/i18n.properties
@@ -168,3 +168,6 @@ TEIID40165=The node tracker failed to communicate with peer nodes.
 TEIID40166=The maximum message size has been exceeded.  The server system property org.teiid.maxMessageSize may need adjusted, or a smaller request may need to be issued.
 
 TEIID40167=Cannot import into non-existent schema {0}.
+
+TEIID40168=Could not create an infinispan cache factory.
+TEIID40169=Could not create an infinispan cache factory.  A default non-concurrent cache will be used instead.  Please consider including the cache-infinispan dependency or manually setting the CacheFactory on the EmbeddedConfiguration.
