+++ wildfly/teiid-feature-pack/wildfly-integration-feature-pack/teiid-releasenotes.html
@@ -47,6 +47,7 @@
 <ul>
   <li><a href="https://issues.jboss.org/browse/TEIID-5476">TEIID-5476</a> JGroups was removed as a direct dependency of the runtime and the associated property removed from the EmbeddedConfiguration.  If you need clustering support with embedded, please raise an issue.</li>
   <li><a href="https://issues.jboss.org/browse/TEIID-5563">TEIID-5563</a> All wildfly specific maven subprojects - including the resource adapter connector-x artifacts - were moved under the org.teiid.wildfly group id.  See the Admin Guide for more migration information.</li>
+  <li><a href="https://issues.jboss.org/browse/TEIID-5563">TEIID-5596</a> The usage of infinispan caching with Teiid Embedded now requires a dependency to org.teiid:cache-infinispan.</li>
 </ul>
 
 <h4>from 11.1</h4>
